chore(ui): Rename hook file
Rename file to align with the naming convention of other files.
